requirements.txt
Flask==0.8
Flask-Login==0.1
Flask-WTF==0.5.2
Jinja2==2.6
WTForms==0.6.3
Werkzeug==0.8.1
pymongo==2.0.1
py-bcrypt==0.2
